+++ tests/Database/Unit/Driver/AbstractDriverTest.php
@@ -8,6 +8,8 @@ use Cycle\Database\Config\SQLiteDriverConfig;
 use Cycle\Database\Driver\Driver;
 use Cycle\Database\Driver\DriverInterface;
 use Cycle\Database\Driver\HandlerInterface;
+use Cycle\Database\Driver\PDOInterface;
+use Cycle\Database\Driver\PDOStatementInterface;
 use Cycle\Database\Query\BuilderInterface;
 use Cycle\Database\Tests\Unit\Stub\TestDriver;
 use Mockery as m;
@@ -87,6 +89,62 @@ class AbstractDriverTest extends TestCase
         $newDriver = clone $driver;
 
         $this->checkImmutability($driver, $newDriver);
+    }
+
+    public function testLogsWithDisabledInterpolation(): void
+    {
+        $pdo = $this->createMock(PDOInterface::class);
+        $pdo
+            ->expects($this->once())
+            ->method('prepare')
+            ->willReturn($this->createMock(PDOStatementInterface::class));
+
+        $driver = TestDriver::createWith(
+            new SQLiteDriverConfig(),
+            $this->createMock(HandlerInterface::class),
+            $this->createMock(BuilderInterface::class),
+            $pdo
+        );
+
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger
+            ->expects($this->once())
+            ->method('info')
+            ->with('SELECT * FROM sample_table WHERE id IN (?, ?, ?) ORDER BY id ASC');
+        $driver->setLogger($logger);
+
+        $driver->query(
+            'SELECT * FROM sample_table WHERE id IN (?, ?, ?) ORDER BY id ASC',
+            [1, 2, 3]
+        );
+    }
+
+    public function testLogsWithEnabledInterpolation(): void
+    {
+        $pdo = $this->createMock(PDOInterface::class);
+        $pdo
+            ->expects($this->once())
+            ->method('prepare')
+            ->willReturn($this->createMock(PDOStatementInterface::class));
+
+        $driver = TestDriver::createWith(
+            new SQLiteDriverConfig(options: ['logQueryParameters' => true]),
+            $this->createMock(HandlerInterface::class),
+            $this->createMock(BuilderInterface::class),
+            $pdo
+        );
+
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger
+            ->expects($this->once())
+            ->method('info')
+            ->with('SELECT * FROM sample_table WHERE id IN (1, 2, 3) ORDER BY id ASC');
+        $driver->setLogger($logger);
+
+        $driver->query(
+            'SELECT * FROM sample_table WHERE id IN (?, ?, ?) ORDER BY id ASC',
+            [1, 2, 3]
+        );
     }
 
     private function checkImmutability(DriverInterface $driver, DriverInterface $newDriver): void